git-svn-id: http://newton-dynamics.googlecode.com/svn@783 b7a2f1d6-d59d-a8fe-1e9e-8d4888b32692

+++ trunk/packages/dCompilerKit/dVirtualMachine/dAssemblerParcer.cpp
@@ -168,8 +168,8 @@ const dAssemblerParcer::dGotoEntry* dAssemblerParcer::FindGoto (const dGotoEntry
 bool dAssemblerParcer::Parce(dAssemblerLexical& scanner)
 {
 	dList<dStackPair> stack;
-	static int actionsCount[] = {2, 2, 5, 1, 5, 5, 2, 2, 2, 2, 5, 5, 5, 5, 5};
-	static int actionsStart[] = {0, 2, 4, 9, 10, 15, 20, 22, 24, 26, 28, 33, 38, 43, 48};
+	static short actionsCount[] = {2, 2, 5, 1, 5, 5, 2, 2, 2, 2, 5, 5, 5, 5, 5};
+	static short actionsStart[] = {0, 2, 4, 9, 10, 15, 20, 22, 24, 26, 28, 33, 38, 43, 48};
 	static dActionEntry actionTable[] = {
 					dActionEntry (40, 0, 1, 0, 0), dActionEntry (256, 0, 4, 0, 0), dActionEntry (40, 0, 1, 0, 0), dActionEntry (256, 0, 4, 0, 0), 
 					dActionEntry (0, 1, 0, 1, 1), dActionEntry (42, 0, 6, 0, 0), dActionEntry (43, 0, 7, 0, 0), dActionEntry (45, 0, 8, 0, 0), 
@@ -186,8 +186,8 @@ bool dAssemblerParcer::Parce(dAssemblerLexical& scanner)
 					dActionEntry (0, 1, 1, 3, 5), dActionEntry (42, 1, 1, 3, 5), dActionEntry (43, 1, 1, 3, 5), dActionEntry (45, 1, 1, 3, 5), 
 					dActionEntry (47, 1, 1, 3, 5)};
 
-	static int gotoCount[] = {2, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0};
-	static int gotoStart[] = {0, 2, 3, 3, 3, 3, 3, 4, 5, 6, 7, 7, 7, 7, 7};
+	static short gotoCount[] = {2, 1, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0};
+	static short gotoStart[] = {0, 2, 3, 3, 3, 3, 3, 4, 5, 6, 7, 7, 7, 7, 7};
 	static dGotoEntry gotoTable[] = {
 					dGotoEntry (258, 2), dGotoEntry (257, 3), dGotoEntry (258, 5), dGotoEntry (258, 11), 
 					dGotoEntry (258, 12), dGotoEntry (258, 13), dGotoEntry (258, 14)};
